test: Refactor and add tests for authentication URL and CIE endpoint
